[dev] lite: fold the branch's tests into one file
Four files with their own copies of the same imports and fixtures became one,
and four cases went: three were CPU duplicates of a GPU case asserting the same
bound with the same parameters, and one checked that a helper is idempotent,
which nothing depends on.

1603 insertions at the start of the review cycle, 976 now.
